Refactor for proper naming of Task classes
